pyproject.toml
[project]
name = "misc-go-grpc-chat"
version = "0.1.0"
description = "A simple gRPC chat app in Go (server) and Python (client)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "grpcio-tools>=1.71.0",
    "grpcio>=1.73.0",
    "protoletariat>=3.3.10",
]

[build-system]
requires = ["setuptools>=65"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = [".", "proto"]
include = ["proto*"]
